removed non-compiling unit tests
