--- radixdlt-java/radixdlt-java/src/test/java/com/radixdlt/client/application/translate/tokens/TransferTokensActionTest.java
@@ -1,68 +1,0 @@
-/*
- * (C) Copyright 2020 Radix DLT Ltd
- *
- * Permission is hereby granted, free of charge, to any person obtaining a
- * copy of this software and associated documentation files (the “Software”),
- * to deal in the Software without restriction, including without limitation
- * the rights to use, copy, modify, merge, publish, distribute, sublicense,
- * and/or sell copies of the Software, and to permit persons to whom the
- * Software is furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be
- * included in all copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED “AS IS”, WITHOUT WARRANTY OF ANY KIND, EXPRESS
- * OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL
- * THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
- * FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
- * DEALINGS IN THE SOFTWARE.
- */
-
-package com.radixdlt.client.application.translate.tokens;
-
-import com.radixdlt.application.TokenUnitConversions;
-import com.radixdlt.identifiers.REAddr;
-import java.math.BigDecimal;
-
-import org.junit.Test;
-
-import com.radixdlt.identifiers.RadixAddress;
-
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static org.mockito.Mockito.mock;
-
-public class TransferTokensActionTest {
-
-	@Test
-	public void testBadBigDecimalScale() {
-		RadixAddress from = mock(RadixAddress.class);
-		RadixAddress to = mock(RadixAddress.class);
-		REAddr rri = mock(REAddr.class);
-
-		BigDecimal tooSmall = BigDecimal.valueOf(1L, TokenUnitConversions.getTokenScale() + 1);
-
-		assertThatThrownBy(() -> TransferTokensAction.create(rri, from, to, tooSmall))
-			.isInstanceOf(IllegalArgumentException.class);
-	}
-
-	@Test
-	public void testSmallestAllowedAmount() {
-		RadixAddress from = mock(RadixAddress.class);
-		RadixAddress to = mock(RadixAddress.class);
-		REAddr rri = mock(REAddr.class);
-
-		assertThat(TransferTokensAction.create(rri, from, to, new BigDecimal("0.00001")).toString()).isNotNull();
-	}
-
-	@Test
-	public void testSmallestAllowedAmountLargeScale() {
-		RadixAddress from = mock(RadixAddress.class);
-		RadixAddress to = mock(RadixAddress.class);
-		REAddr rri = mock(REAddr.class);
-
-		assertThat(TransferTokensAction.create(rri, from, to, new BigDecimal("0.000010000")).toString()).isNotNull();
-	}
-}
