--- radixdlt-java/radixdlt-java/src/test/java/com/radixdlt/client/util/MagicByteTest.java
@@ -1,40 +1,0 @@
-/*
- * (C) Copyright 2020 Radix DLT Ltd
- *
- * Permission is hereby granted, free of charge, to any person obtaining a
- * copy of this software and associated documentation files (the “Software”),
- * to deal in the Software without restriction, including without limitation
- * the rights to use, copy, modify, merge, publish, distribute, sublicense,
- * and/or sell copies of the Software, and to permit persons to whom the
- * Software is furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be
- * included in all copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED “AS IS”, WITHOUT WARRANTY OF ANY KIND, EXPRESS
- * OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL
- * THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
- * FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
- * DEALINGS IN THE SOFTWARE.
- */
-
-package com.radixdlt.client.util;
-
-import org.junit.Test;
-
-import static junit.framework.TestCase.assertEquals;
-
-public class MagicByteTest {
-
-    @Test
-    public void testMagicByte() {
-        int magicIntAlphanet = 281411585;
-        byte magicByte = (byte) (magicIntAlphanet & 0xff);
-        /// The origin of this byte value is this library it self, commit: acbc5307cf5c9f7e1c30300f7438ef5dbc3bb629
-        /// This value can be used as a reference for other Radix libraries, e.g. Swift.
-        byte expected = (byte) 1;
-        assertEquals(expected, magicByte);
-    }
-}
